Refactor database handling and update Docker configuration
