+++ Dockerfile
@@ -7,4 +7,4 @@ RUN pip install --no-cache-dir -r requirements.txt
 
 COPY application/ .
 
-CMD ["python", "app.py"]
+CMD ["flask", "--app", "app", "run"]
