--- docker/.env
@@ -1,11 +1,0 @@
-DB_HOST = postgres
-DB_NAME = scoresdb
-DB_USER = admin
-DB_PASSWORD = admin1234
-
-POSTGRES_USER = admin
-POSTGRES_PASSWORD = admin1234
-POSTGRES_DB = scoresdb
-
-REDIS_HOST = redis
-REDIS_PORT = 6379
