--- docker/flask-app-compose.yml
@@ -1,18 +1,0 @@
-version: "3.9"
-
-services:
-  flask-app:
-    build:
-      context: ..
-    container_name: flask_app_production
-    depends_on:
-      - postgres
-    environment:
-      DB_HOST: ${DB_HOST}       
-      DB_NAME: ${DB_NAME}        
-      DB_USER: ${DB_USER}          
-      DB_PASSWORD: ${DB_PASSWORD} 
-      REDIS_HOST: ${REDIS_HOST}
-      REDIS_PORT: ${REDIS_PORT}
-    ports:
-      - "5000:5000"
