--- docker/postgres-compose.yml
@@ -1,17 +1,0 @@
-version: "3.9"
-
-services:
-  postgres:
-    image: postgres:15
-    container_name: postgres_production
-    environment:
-      POSTGRES_USER: ${POSTGRES_USER}
-      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
-      POSTGRES_DB: ${POSTGRES_DB}
-    ports:
-      - "5432:5432"
-    volumes:
-      - postgres_data:/var/lib/postgresql/data
-
-volumes:
-  postgres_data:
