--- docker/redis-compose.yml
@@ -1,8 +1,0 @@
-version: "3.9"
-
-services:
-  redis:
-    image: redis:7
-    container_name: redis_container
-    ports:
-      - "6379:6379"
